+++ Repository/Repositories/TeacherRepository.cs
@@ -1,4 +1,5 @@
 ﻿using Repository.Entities;
+using Repository.Entities.Enums;
 using Repository.Interfaces;
 using System;
 using System.Collections.Generic;
@@ -21,14 +22,15 @@ namespace Repository.Repositories
 			this.context.Teachers.Add(item);
 			this.context.Save();
 			return item;
-		}
-
-		
+		}	
 		public override Teacher DeleteItem(string id)
 		{
 			Teacher item = GetById(id);
-			this.context.Teachers.Remove(item);
-			this.context.Save();
+			if(item !=null)
+			{
+                this.context.Teachers.Remove(item);
+                this.context.Save();
+            }		
 			return item;
 		}
 
@@ -44,18 +46,20 @@ namespace Repository.Repositories
 			return this.context.Teachers.FirstOrDefault(t => t.Id.Equals(id));
 		}
 
-		
-
 		public override Teacher UpdateItem(string id, Teacher item)
 		{
 			Teacher teacher = GetById(id);
-			teacher.Password = item.Password;
-			teacher.Name = item.Name;
-			teacher.Address = item.Address;
-			teacher.DateOfBirth = item.DateOfBirth;
-			teacher.Classes = item.Classes;
-			teacher.Role = item.Role;
-			teacher.Phone = item.Phone;
+
+            if (teacher == null)
+				return null;
+
+            teacher.Password = item.Password != null ? item.Password : teacher.Password;
+			teacher.Name = item.Name != null ? item.Name : teacher.Name;
+			teacher.Address = item.Address != null ? item.Address : teacher.Address;
+			teacher.DateOfBirth = item.DateOfBirth != null ? item.DateOfBirth : teacher.DateOfBirth;
+			teacher.Classes = item.Classes != null ? item.Classes : teacher.Classes;
+			teacher.Role = item.Role != Roles.None ? item.Role : teacher.Role;
+			teacher.Phone = item.Phone != null ? item.Phone : teacher.Phone;
 
 			this.context.Save();
 			return teacher;
